1.0.0 Out of Beta!
Wow! This has been a big project! This update includes tons of changes, new features and bug fixes! This update goes from the bottom from some grammar fixes to pretty much redoing the whole admin folder! The bot is now officially out of Beta and should now be stable and usable. More updates are to come with new features, bug fixes, and changes to the commands and features. A complete changelog can be found in the Changelogs folder then 1.0.0+ File then look for the 1.0.0 update! More awesome updates are soon to come! Sincerely, theopcoder